+++ NEWS.md
@@ -1,6 +1,8 @@
 # jacobi 2.2.0
 
-Some values of the Jacobi theta functions were wrong as of version 2.1.0.
+- Some values of the Jacobi theta functions were wrong as of version 2.1.0.
+
+- Added some unit tests.
 
 
 # jacobi 2.1.0